feat(form): User's can create forms

+++ src/main/java/com/jivecake/api/request/EntityQuantity.java
@@ -2,5 +2,5 @@ package com.jivecake.api.request;
 
 public class EntityQuantity<A> {
     public A entity;
-    public int quantity;
+    public long quantity;
 }
